requirements.txt
python-dotenv
openai
numpy
torch
sentence-transformers
faiss